fix(parser): Fix double image downloading
There were situations in which you could have what was almost a phantom
modal pop up and the XPATH query that I used to select the elements
would get the images from the phantom modal and the actual model of a
daily report. In this case, in addition to the regular images for a day,
you get another copy of the same images every single day. The solution
was to improve the XPATH query so it forced everything to be in the
visible modal. However, this might be fragile because XPATH selectors do
not have knowledge of the multi-valued class field (unlike CSS
selectors), so you need to have an exact match on the class.

DCO-1.1-Signed-off-by: Patrick Wagstrom <patrick@wagstrom.net>