feat(exif): first version of exif tagger program
In the original version of this application, I had both the scraper and
exif tagging tied into one program. This resulted in something that was
really fragile. This starts to bring exif tagging back into the program
by adding in a program that will read the output log and go over the
images to tag them with the appropriate exif data so all the images show
up nicely when imported into programs like Apple Photos.

DCO-1.1-Signed-off-by: Patrick Wagstrom <patrick@wagstrom.net>